requirements.txt
Pillow
rich
requests
pathlib